backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.28.0
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.1
setuptools==70.0.0
simplejson==3.19.2
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
